Stop FH immediately when running out of entries. Possibility to set first entry at command line

+++ postProcessing2/source/runFH.cxx
@@ -1,4 +1,3 @@
-
 #include "slhaea.h"
 #include <iostream>
 #include <fstream>
@@ -15,49 +14,69 @@
 int main(int arc, char** argv){
 
   std::string model = argv[1];
-  std::string numberOfEventsStr = argv[2];
+  std::string firstEventStr = argv[2];
+  std::string numberOfEventsStr = argv[3];
 
-  std::stringstream numberOfEventsStream;
-  int numberOfEvents;
-  numberOfEventsStream<<numberOfEventsStr;
-  numberOfEventsStream>>numberOfEvents;
 
+  std::stringstream sstream;
+  
+  Long64_t numberOfEvents;
+  sstream.str(numberOfEventsStr);
+  sstream>>numberOfEvents;
+  sstream.clear();
+
+  Long64_t firstEvent;
+  sstream.str(firstEventStr);
+  sstream>>firstEvent;
+  sstream.clear();
 
   TString ifDir="$FITTINO_INTERFACE/FH_noHiggsRates/"+model;
-  TString lockDir="$FITTINO_LOCKING/FH_noHiggsRates/"+model;
-
   gSystem->ExpandPathName(ifDir);
+  TString lockDir="$FITTINO_LOCKING/FH_noHiggsRates/"+model;
   gSystem->ExpandPathName(lockDir);
 
+  bool useInterFaceFile=false;
+  if ( firstEvent<0 ) useInterFaceFile=true;
+
   int rc=1;
-  while (rc!=0){
-    rc=system(("mkdir "+lockDir+"/locked").Data());
+
+  if (useInterFaceFile){
+  
+    while (rc!=0){
+      rc=system(("mkdir "+lockDir+"/locked").Data());
+    }
+  
+    ifstream fin((ifDir+"/event.txt").Data());
+    fin >> firstEventStr;
+    fin.close();
+
+    sstream.str(firstEventStr);
+    sstream>>firstEvent;
+    sstream.clear();
+
   }
+
+  Long64_t lastEvent = firstEvent+numberOfEvents;
+
+  std::string lastEventStr;
+  sstream.str(std::string());
+  sstream << lastEvent;
+  lastEventStr=sstream.str();
+  sstream.clear();
+
+  if (useInterFaceFile){
+
+    system(("mv "+ifDir+"/event.txt.last "+ifDir+"/event.txt.last2").Data()); 
+    system(("mv "+ifDir+"/event.txt "+ifDir+"/event.txt.last").Data()); 
+
+    ofstream ifFile;
+    ifFile.open((ifDir+"/event.txt").Data());
+    ifFile << lastEvent<< std::endl;
+    ifFile.close();
   
-  ifstream fin((ifDir+"/event.txt").Data());
-  int firstEvent;
-  fin >> firstEvent;
-  fin.close();
+    system(("rm -r "+lockDir+"/locked").Data());
 
-  system(("mv "+ifDir+"/event.txt.last "+ifDir+"/event.txt.last2").Data()); 
-  system(("mv "+ifDir+"/event.txt "+ifDir+"/event.txt.last").Data()); 
-
-  int lastEvent = firstEvent+numberOfEvents;
-
-  ofstream ifFile;
-  ifFile.open((ifDir+"/event.txt").Data());
-  ifFile << lastEvent<< std::endl;
-  ifFile.close();
-  
-  system(("rm -r "+lockDir+"/locked").Data());
-
-  std::stringstream firstEventStream; 
-  firstEventStream << firstEvent;
-  std::string firstEventStr=firstEventStream.str();
-
-  std::stringstream lastEventStream; 
-  lastEventStream << lastEvent;
-  std::string lastEventStr=lastEventStream.str();
+  }
 
   TChain* markovChain_in = new TChain("markovChain");
   markovChain_in->Add(("$FITTINO_SORT/outputs/"+model+"/*.root").c_str());
@@ -67,7 +86,12 @@ int main(int arc, char** argv){
   metadata_in->Add(("$FITTINO_SORT/outputs/"+model+"/*.root").c_str());
   metadata_in->LoadTree(0);   
 
-  TFile* outFile= new TFile((model+"_evnt_"+firstEventStr+"_"+lastEventStr+".root").c_str() , "RECREATE");
+
+  std::string outFileName=model+"_evnt_"+firstEventStr+"_"+lastEventStr+".root";
+
+  std::cout<<"Creating out file "<<outFileName<<std::endl;
+
+  TFile* outFile= new TFile(outFileName.c_str() , "RECREATE");
   
   TTree *metadata_out = metadata_in->GetTree()->CloneTree(0); 
 
@@ -83,10 +107,20 @@ int main(int arc, char** argv){
   FHCalculator fh;
   fh.SetOutput(&data);  
 
-  for (int i=firstEvent; i<lastEvent; i++){
+  std::cout<<"Looping from "<<firstEvent<<" to "<<lastEvent<<std::endl; 
 
-    markovChain_in->GetEntry(i);
-    metadata_in->GetEntry(i);
+  for (Long64_t i=firstEvent; i<lastEvent; i++){
+
+    int nb1=markovChain_in->GetEntry(i);
+    int nb2=metadata_in->GetEntry(i);
+
+    if (nb1<=0 || nb2<=0){
+      
+      std::cout<<"Problem reading entry "<<i<<std::endl;
+      std::cout<<"Maybe end of chain reached. Stopping. "<<std::endl;
+      break;
+      
+    }
 
     rc=spheno.Calculate();
     if (rc) continue;
